requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

python_telegram_bot == 13.9
